requirements.txt
google-generativeai
PyPDF2
streamlit


#os
